Revert "corosync: handle network partition"
This reverts commit 15df161958a38cf3f7bc83b5bc2c8a1817b3072e.

The commit introduces a better handling method of network
partition. But the method seems to have a possibility of losing event
from corosync. According to the experiment by Tsuji-san, dispatching
of leave event can be delayed because of the poll(2) in
__corosync_dispatch(). I think removing the poll(2) would be safer
because corosync doesn't guarantee it calls callback function
immediately if the fd of cpg has unread data.

Reported bug: https://bugs.launchpad.net/sheepdog-project/+bug/1368503

Conflicts:
	sheep/cluster/corosync.c
[ Conflicts were resolved by Hitoshi Mitake ]

Reported-by: Masahiro Tsuji <tuji@atworks.co.jp>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>