Revert "corosync: handle network partition"
This reverts commit 15df161958a38cf3f7bc83b5bc2c8a1817b3072e.

The commit introduced a better handling method of network
partition. But the method seems to have a possibility of losing event
from corosync. According to an experiment by Tsuji-san, dispatching
of leave event can be delayed because of the poll(2) in
__corosync_dispatch() with corosync < 2.x.

In the next commit, simpler method for detecting network partition
based on a parameter of cpg callback function will be implemented. The
reverted method based on poll(2) depends on semantics of fd obtained
by cpg_fd_get(3) of corosync. The poll(2) based method worked well on
corosync 1.x, but it doesn't work well on 2.x. Internal big change
of corosync (maybe employing libqb) seemed to change the semantics so
we cannot use poll(2) for detecting queued messages.

With this change, sheepdog cannot use corosync 1.x as its cluster
manager any more. But the version has a critical bug in its message
delivery mechanism which can produce broken epoch log. So it is
reasonable to force users to stop using the version and introduce
this change.

Reported bug: https://bugs.launchpad.net/sheepdog-project/+bug/1368503

Conflicts:
	sheep/cluster/corosync.c
[ Conflicts were resolved by Hitoshi Mitake ]

Reported-by: Masahiro Tsuji <tuji@atworks.co.jp>
Tested-by: Masahiro Tsuji <tuji@atworks.co.jp>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>

Conflicts:

	sheep/cluster/corosync.c